requirements.txt
PIL==1.1.7
SimpleCV==1.3
beautifulsoup4==4.3.2
clime==0.2.7
lxml==2.3.2
mock==1.0.1
nose==1.1.2
numpy==1.6.1
psycopg2==2.5.4
pygeocoder==1.2.5
python-sql==0.6
scikit-image==0.10.1
scipy==0.9.0
simplejson==2.3.2
titlecase==0.7.2
unittest2==0.8.0
